Streamlit_Tutorial/pyproject.toml
[tool.poetry]
name = "streamlit-tutorial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.34.0"
python-dotenv = "^1.0.1"
openai = "^1.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
